upgrade to the latest Django, v. 1.6
Dormbase was previously using Django 1.4, and it wasn't too hard to switch.
(The main problem was a difference in the url function in views, but a quick
sed fixed that.)

I'm not entirely sure what the benefit of upgrading is because I'm new to
Django, but I think it's better to upgrade now (before we have a giant
codebase) than later.