remove all search functionality
As I'm constantly updating model code, it doesn't make sense to have hard-coded
in search indices. We can reconsider later whether we want to add Haystack
back, though I'm skeptical -- all the use cases I can think of for search are a
permissions-restricted (ie, to just one site) search on a single model, for
which a simple HTML form with a Django model filter suffices.