add django-guardian
We're going to want site-based permissions (eg, User is a deskworker *for this
one dorm* only), and it seems like the easiest way to do so is django-guardian,
which lets you specify permissions for a specific instance of an object. We'll
use the convention whereby the only objects we pass to guardian are site
objects, so effectively this lets us use permissions on a site-by-site basis.