Use South and fixtures to add Resident permissions
Wow. This is ugly, but I really can't find a better way.

We need some sort of permission system. Furthermore, we need to attach
permissions to dorms. (Eg, an AD whose job covers two dorms should have
certain permissions on those two dorms but none others. So immediately,
we can't just check in code where they live because a user might have
permissions for multiple dorms.)

Permissions do seems like the right route to take. I've decided to use
django-guardian to couple permissions to Site objects. We can then
assign permissions to users by creating groups (eg, Random Hall
deskworkers), putting users in groups, and assigning permissions to
these groups. This lets each dorm individually (via the admin console)
decide what deskworkers should be able to do, rather than enforcing a
universal set of rules in code.

The biggest catch is that django-guardian requires us to associate the
permissions with the site app name, but (1) we can't edit the Site class
directly, and (2) even if we could it violates modularity to have a list
of all permissions in Site. The list of permissions for each app must
live in that app.

With that in mind, the best solution I found was to create a fixture for
each app with the permissions it needs to create, and then to set up
South to install these fixtures.

As an example I created a can_modify_residents permission living in the
residents app. Apparently the South datamigration has JSON in it
describing the Resident table and all joined tables. If the schema
changes before this code goes into production, we may want to
re-generate these migrations.