move some templates
I realized that a couple templates were probably *not* changing on a
dorm-by-dorm basis, so we should move these up a level. (Of course, a
dorm can still override them if the dorm wants.)

Also, we no longer need a logout template because we've set logout to
always redirect to the home view.

Finally, I realized that we need both SITE_DIR and BASE_DIR in the
static/template search paths so that Django searches SITE_DIR before
BASE_DIR. (Otherwise it still searches SITE_DIR, since it's part of an
app, but only after searching BASE_DIR, which is backwards.)