set up templates and views for the personal sub-pages
The personal section of the site is split into five subpages which you
can access by clicking personal and then navigating using tabs.

This may not be the best UX choice for the long run (I don't quite like
the dropdown style of the facilities navbar item, and I couldn't find a
simple way to load tabs with AJAX in Bootstrap), but it works for now.