add a profile page and a resident_required decorator
The profile page is for dorm residents only and shows a read-only
profile. In the future, this may be extended to show more info and/or be
editable.

The resident_required decorator is like a login_required decorator, but
also checks that the user is a resident of a specified dorm.

When a non-resident tries to visit the personal section, they now see a
brief notice.