add deskworker permissions
Adds permissions (via South-managed fixtures) for:
-viewing the desk site
-adding deskworkers
-managing guest lists
-managing packages
-managing desk items

The reason we're manually specifiying primary keys is so that, in a
later fixture, we can grant these permissions to the deskworker group.
(This is easier than doing so in /admin because I'm wiping the database
a lot while testing.) I'm starting pk at 100 because 1-15 or so are
taken and I really want to avoid any potential conflicts.

Now, I know, it's horrible to hard-code in and use IDs. Especially when
it's actually hard-coded and not a variable. (JSONs don't take
variables, ugh.) *But* I claim that this isn't actually making things
worse -- because if we didn't hard-code in primary keys, we'd need our
permission-granting fixture to use the codenames, which are also
arbitrary and not stored in a variable. (Why can't django-guardian
just have natural keys, why.)

So, yeah, it's ugly, but not that bad.