add fixture to assign permissions to the deskworker group
I really don't like that I need to hardcode in a bunch of things, but
it's inevitable. The hardcoded items are:
-permission IDs
-site ID
-site name
-deskworker group name

Luckily, these are all deterministically set in our system, so there's a
100% chance that things will work as of now. The only way things can
break is if someone modifies one of these values in a different file and
doesn't update the reference here... which is definitely bound to happen
at some point.