deskworker interface for package pickup
You can see a table of all packages, and you simply click a link when a
package is picked up. I wanted to make the common case of picking up
multiple packages easy (ie, you search for the resident's name, then
need to mark multiple packages as retrieved). My best solution from a UX
standpoint was to have the retrieved package still show in the table
(for an hour) and have the pick up link fade out. The table will even
stay on the same page! The only weird corner case is if you have the
table sorted by pick-up status; then, after the fade-out, the rows shift
up one..